# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo-list API documentation",
        "version": "1.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registration successfull"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Authenticate user and generate token",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registration successfull"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Revoque the authenticated user",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "User revoqued successfully"
                    }
                }
            }
        },
        "/api/todo/create": {
            "post": {
                "tags": [
                    "Todo"
                ],
                "summary": "Creates a new task",
                "operationId": "935868959ad2648d6397068064907e8b",
                "parameters": [
                    {
                        "name": "title",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "path",
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "path",
                        "schema": {
                            "type": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/todo/get": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Get task information",
                "operationId": "36e871c94e327472dd8a665970e11e1f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task retrieved successfully"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/todo/remove": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Get task information",
                "operationId": "6a0c5335e9d049ca7eee01fdd21c1f63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task removed successfully"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get authenicated user information",
                "operationId": "e85672f7ffaba7b2e129a7b7f02194a5",
                "responses": {
                    "200": {
                        "description": "User information retrieved"
                    }
                }
            }
        },
        "/api/user/list": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Retrieve user task list",
                "operationId": "190f236e115a2794c767ba30f3a94b47",
                "responses": {
                    "200": {
                        "description": "List retrieved"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}